W3-D6-MLFlow-DVC/requirements.txt
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.1
joblib==1.4.2
pytest==8.3.2
pyyaml==6.0.2
matplotlib==3.8.4
mlflow==2.14.1
dvc[s3]==3.51.2
